Fix: Remove extra files and rename supabase
