+++ examples/react/start-supabase-basic/.env
@@ -1,9 +1,2 @@
-# Environment variables declared in this file are automatically made available to Prisma.
-# See the documentation for more detail: https://pris.ly/d/prisma-schema#accessing-environment-variables-from-the-schema
-
-# Prisma supports the native connection string format for PostgreSQL, MySQL, SQLite, SQL Server, MongoDB and CockroachDB.
-# See the documentation for all the connection string options: https://pris.ly/d/connection-strings
-
-DATABASE_URL="file:./dev.db"
 SUPABASE_URL=PleaseChangeMe
 SUPABASE_ANON_KEY=PleaseChangeMe
