--- examples/react/start-supabase-basic/app/utils/prisma.ts
@@ -1,16 +1,0 @@
-import crypto from 'node:crypto'
-import { PrismaClient } from '@prisma/client'
-
-export const prismaClient = new PrismaClient()
-
-export function hashPassword(password: string) {
-  return new Promise<string>((resolve, reject) => {
-    crypto.pbkdf2(password, 'salt', 100000, 64, 'sha256', (err, derivedKey) => {
-      if (err) {
-        reject(err)
-      } else {
-        resolve(derivedKey.toString('hex'))
-      }
-    })
-  })
-}
